requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
requests==2.31.0
